--- source/IrcDotNet/project.json
@@ -1,27 +1,0 @@
-﻿{
-  "version": "0.6.1",
-  "releaseNotes": "https://github.com/IrcDotNet/IrcDotNet/releases/tag/0.6.0",
-
-  "summary": "A complete IRC (Internet Relay Chat) client library for the .NET Framework",
-  "description": "IRC.NET aims to provide a complete and efficient implementation of the protocol as described in RFCs 1459 and 2812, as well as de-facto modern features of the protocol.",
-  "copyright": "IRC.NET is copyright © 2011-2016 Alex Regueiro, Christian Stewart",
-  "authors": [ "Alexander Regueiro", "Christian Stewart" ],
-  "owners": [ "Alexander Regueiro", "Christian Stewart" ],
-  "tags": [ "communication", "irc", "networking", "ctcp" ],
-
-  "projectUrl": "https://IrcDotNet.github.io/IrcDotNet/",
-  "iconUrl": "https://raw.githubusercontent.com/IrcDotNet/IrcDotNet/master/doc/content/img/logo.png",
-
-  "licenseUrl": "https://github.com/IrcDotNet/IrcDotNet/License.txt",
-  "requireLicenseAcceptance": true,
-
-  "frameworks": {
-    "net40": { },
-    "net45": { },
-    "net451": { },
-    "dnx451": { }
-  },
-
-  "dependencies": {
-  }
-}
